Refactor the last IndexHandler into view_base.py.
